feat(compile): install local dependencies
If a Gemfile is present in the project root, we can assume that
the user would like to install those dependencies. This change
installs bundler and installs the dependencies required. If no
Gemfile is present, no action is taken, so it's a no-op for
apps that do not have a Gemfile.

fixes #2